Links up checkpoints to the navigation computer complete with distance to waypoints, waypoint number, and waypoint name.

+++ Canyon Run/Assets/Scripts/Conversions.cs
@@ -24,7 +24,7 @@ public static class Conversions
     {
         return Math.Round(speedInMetersPerSecond * 3.28084f);
     }
-    public static double MeterstoNauticalMiles(float distanceInMeters)
+    public static double MeterstoNauticalMiles(double distanceInMeters)
     {
         return Math.Round(distanceInMeters / 1852, 1);
     }
